Add files via upload
You are required to compute the power of a number by implementing a calculator.Create a class MyCalculator which consists of a single method power(int, int). This method takes two integers, and , as parameters and finds . If either or is negative, then the method must throw an exception which says "n and p should be non-negative".
Please read the partially completed code in the editor and complete it. Your class mustn't be public. No need to worry about constraints, there won't be any overflow if your code is correct.